requirements.txt
click==8.0.1
colorama==0.4.4
colorlog==5.0.1
colormap==1.0.3
cycler==0.10.0
easydev==0.11.1
Flask==2.0.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.21.1
pexpect==4.8.0
Pillow==8.3.1
ptyprocess==0.7.0
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
Werkzeug==2.0.1
